refactor(components): decompose layout in Order
